pyproject.toml
[project]
name = "todo-app"
version = "2.0.0"
description = "Evolution of Todo - Phase 2: SQLModel + Database Persistence"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.12.5",
    "rich>=14.2.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--verbose --cov=src --cov-report=term-missing"
